Swap testgen workflow
